% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{AIC.lrgpr}
\alias{AIC.lrgpr}
\title{Akaike's Information Criterion (AIC)}
\usage{
AIC.lrgpr(object, ..., k = 2)
}
\arguments{
\item{object}{model fit with \code{\link{lrgpr}}}

\item{...}{other arguments}

\item{k}{for compotability, not used}
}
\description{
AIC for model fit by \code{\link{lrgpr}}
}
