% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{BIC.lrgpr}
\alias{BIC.lrgpr}
\title{Bayesian Information Criterion (BIC)}
\usage{
\method{BIC}{lrgpr}(object, ...)
}
\arguments{
\item{object}{model fit with \code{\link{lrgpr}}}

\item{...}{other arguments}
}
\description{
BIC for model fit by \code{\link{lrgpr}}
}
