% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{coefficients.lrgpr}
\alias{coefficients.lrgpr}
\title{Extract Model Coefficients}
\usage{
\method{coefficients}{lrgpr}(object, ...)
}
\arguments{
\item{object}{model fit with \code{\link{lrgpr}}}

\item{...}{other arguments}
}
\description{
Coefficients estimated with \code{\link{lrgpr}}
}
