% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrgpr.R
\name{convertToBinary}
\alias{convertToBinary}
\title{Convert ASCII to binary file}
\usage{
convertToBinary(
  filename,
  filenameOut,
  format,
  nthreads = detectCores(logical = TRUE),
  onlyCheckFormat = FALSE,
  rowNames = NULL,
  simpleAnnotation = TRUE
)
}
\arguments{
\item{filename}{file to be converted}

\item{filenameOut}{name of binary file produced}

\item{format}{specify `TPED', `DOSAGE' or `GEN'}

\item{nthreads}{number of threads to use}

\item{onlyCheckFormat}{only check the format of the input file and don't perform conversion}

\item{rowNames}{Define custom rowNames}

\item{simpleAnnotation}{write FILE_alleles with name, allele1, allele2.  If FALSE, write chrom id genetic_position position allele1 allele2}
}
\description{
Converts TPED/DOSAGE/GEN files to binary format
}
\details{
\itemize{
\item{TPED:	}{plink file can be in either --recode or --recode12 format}
\item{DOSAGE:	}{file follows plink format: http://pngu.mgh.harvard.edu/~purcell/plink/dosage.shtml}
}

Example:

    SNP    A1  A2   F1   I1     F2   I2     F3   I3

  rs0001   A   C   0.98 0.02   1.00 0.00    0.00 0.01

  rs0002   G   A   0.00 1.00   0.00 0.00    0.99 0.01

where the F* values correspond to the dosage values

\itemize{
\item{GEN:	}{file follows OXFORD format}
}
}
