\name{error.bar}
\alias{error.bar}
\title{Plot Error Bars}
\usage{
  error.bar(x, y, upper, lower = upper, length = 0.1, ...)
}
\arguments{
  \item{x}{x-axis position}

  \item{y}{y-axis position}

  \item{upper}{height of bar above y}

  \item{lower}{height of bar below y}

  \item{length}{horizontal length of the error bar}

  \item{...}{arguments for arrows(...)}
}
\description{
  Plot error bars for a confidence interval
}

