% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrgpr.R
\name{getMACHrsq}
\alias{getMACHrsq}
\title{Evaluate MACH r^2 information metric for each column}
\usage{
getMACHrsq(X, nthreads = detectCores(logical = TRUE), progress = TRUE)
}
\arguments{
\item{X}{matrix where each column is a marker}

\item{nthreads}{number of threads to use}

\item{progress}{show progress bar 

See definition in Supplementary information S3 for Marchini and Howie (2010): \url{http://www.nature.com/nrg/journal/v11/n7/extref/nrg2796-s3.pdf}.

#' @references
 Marchini, J. and B. Howie. (2010) Genotype imputation for genome-wide association studies. _Nature Reviews Genetics_ 11, 499-511
'}
}
\description{
Evaluate MACH r^2 information metric for each column
}
