% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrgpr.R
\name{glmApply2}
\alias{glmApply2}
\title{(Experimental) faster version of glmApply}
\usage{
glmApply2(
  formula,
  features,
  terms = NULL,
  family = gaussian(),
  useMean = TRUE,
  nthreads = detectCores(logical = TRUE),
  verbose = FALSE,
  progress = TRUE,
  cincl = c(),
  cexcl = c()
)
}
\description{
Like glmApply, by linear instead of quadratic as a function of the number of covariates.  This is still experimental
}
