% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{influence.lrgpr}
\alias{influence.lrgpr}
\title{Regression Diagnostics}
\usage{
\method{influence}{lrgpr}(model, ...)
}
\arguments{
\item{model}{model fit with \code{\link{lrgpr}}}

\item{...}{other arguments}
}
\description{
Basic quantities for regression diagnostics from fit of \code{\link{lrgpr}}
}
