% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{leverage.lrgpr}
\alias{leverage.lrgpr}
\title{Regression Diagnostics}
\usage{
leverage.lrgpr(object)
}
\arguments{
\item{object}{model fit with \code{\link{lrgpr}}}
}
\description{
Basic quantities for regression diagnostics from fit of \code{\link{lrgpr}}
}
