% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{lm.influence.lrgpr}
\alias{lm.influence.lrgpr}
\title{Regression Diagnostics}
\usage{
lm.influence.lrgpr(object, ...)
}
\arguments{
\item{object}{model fit with \code{\link{lrgpr}}}

\item{...}{other arguments}
}
\description{
Basic quantities for regression diagnostics from fit of \code{\link{lrgpr}}
}
