\name{plot.criterion.lrgpr}
\alias{plot.criterion.lrgpr}
\title{Plot AIC/BIC/GCV values for lrgpr() model as rank changes}
\usage{
  plot.criterion.lrgpr(x, col = rainbow(3), ...)
}
\arguments{
  \item{x}{list returned by 'criterion.lrgpr'}

  \item{col}{array of 3 colors}

  \item{...}{}
}
\description{
  'plot.criterion.lrgpr' plots the criteria returned by
  'criterion.lrgpr'
}
\seealso{
  criterion.lrgpr
}

