% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{print.lrgpr}
\alias{print.lrgpr}
\title{Print Values}
\usage{
\method{print}{lrgpr}(x, ...)
}
\arguments{
\item{x}{model fit from \code{\link{lrgpr}}}

\item{...}{other arguments}
}
\description{
Print details for fit from \code{\link{lrgpr}}
}
