% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{print.summary.lrgpr}
\alias{print.summary.lrgpr}
\title{Object Summaries}
\usage{
\method{print}{summary.lrgpr}(x, ...)
}
\arguments{
\item{x}{model fit from \code{\link{lrgpr}}}

\item{...}{other arguments}
}
\description{
Print summary for fit from \code{\link{lrgpr}}
}
