% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{read.tfam}
\alias{read.tfam}
\title{Read plink FAM/TFAM files}
\usage{
read.tfam(file, checkFileExtension = TRUE)
}
\arguments{
\item{file}{location of FAM/TFAM file}

\item{checkFileExtension}{If TRUE, throw error if file doesn't end with fam/tfam}
}
\description{
Read FAM/TFAM file into a dataframe.  This function is the same as read.fam
}
