\name{residuals.lrgpr}
\alias{residuals.lrgpr}
\title{Extract Model Residuals}
\usage{
  residuals.lrgpr(object, type = "working", ...)
}
\arguments{
  \item{object}{model fit with \code{\link{lrgpr}}}

  \item{type}{the type of residual, but there is only one
  option here}

  \item{...}{other arguments}
}
\description{
  Residuals fitted with \code{\link{lrgpr}}
}

