\name{rstandard.lrgpr}
\alias{rstandard.lrgpr}
\title{Regression Deletion Diagnostics}
\usage{
  rstandard.lrgpr(model, ...)
}
\arguments{
  \item{model}{model fit with \code{\link{lrgpr}}}

  \item{...}{other arguments}
}
\description{
  Basic quantities for regression deletion diagnostics from
  fit of \code{\link{lrgpr}}
}

