% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{rstandard.lrgpr}
\alias{rstandard.lrgpr}
\title{Regression Deletion Diagnostics}
\usage{
\method{rstandard}{lrgpr}(model, ...)
}
\arguments{
\item{model}{model fit with \code{\link{lrgpr}}}

\item{...}{other arguments}
}
\description{
Basic quantities for regression deletion diagnostics from fit of \code{\link{lrgpr}}
}
