% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{summary.lrgpr}
\alias{summary.lrgpr}
\title{Summarizing LRGPR / Linear Mixed Model Fits}
\usage{
\method{summary}{lrgpr}(object, ...)
}
\arguments{
\item{object}{model fit from \code{\link{lrgpr}}}

\item{...}{other arguments}
}
\description{
Print summary for fit from \code{\link{lrgpr}}
}
