\name{vcov.lrgpr}
\alias{vcov.lrgpr}
\title{Calculate Variance-Covariance Matrix for a \code{\link{lrgpr}} Object}
\usage{
  vcov.lrgpr(object, ...)
}
\arguments{
  \item{object}{model fit with \code{\link{lrgpr}}}

  \item{...}{}
}
\description{
  Returns the variance-covariance matrix of the main
  parameters of a fitted model object
}

