\name{convertToBinary}
\alias{convertToBinary}
\title{Convert ASCII to binary file}
\usage{
  convertToBinary(filename, filenameOut, format)
}
\arguments{
  \item{filename}{file to be converted}

  \item{filenameOut}{name of binary file produced}

  \item{format}{specify "TPED" or "DOSAGE"}
}
\description{
  'convertToBinary' converts TPED and DOSAGE files to
  binary format
}

