% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getAlleleFreq}
\alias{getAlleleFreq}
\title{Calculate allele frequency}
\usage{
getAlleleFreq(X, nthreads = detectCores(logical = TRUE), progress = TRUE)
}
\arguments{
\item{X}{matrix where each column is a marker coded 0,1,2 or with dosage values in this range}

\item{nthreads}{number of threads to use}

\item{progress}{show progress bar}
}
\description{
Calculate allele frequency
}

