% Generated by roxygen2 (4.0.1): do not edit by hand
\name{glmApply}
\alias{glmApply}
\title{Fit standard (i.e. fixed effects) linear or logistic model for many markers}
\usage{
glmApply(formula, features, terms = NULL, family = gaussian(),
  useMean = TRUE, nthreads = detectCores(logical = TRUE), verbose = FALSE,
  progress = TRUE, cincl = c(), cexcl = c())
}
\arguments{
\item{formula}{standard linear modeling syntax as used in `lm'.  SNP is a place holder for the each successive column of features}

\item{features}{a matrix where the statistical model is evaluated with SNP if formula replace by each column successively}

\item{terms}{indices of the coefficients to be tested.  The indices corresponding to SNP are used if terms is not specified}

\item{family}{gaussian() for a continuous response, and binomial() to fit a logit model for a binary response}

\item{useMean}{if TRUE, replace missing entries with column mean.  Otherwise, do not evaluate the model for that column}

\item{nthreads}{number of to use for parallel execution}

\item{verbose}{print additional information}

\item{progress}{show progress bar}

\item{cincl}{column indeces of features to include for analysis}

\item{cexcl}{column indeces of features to exclude for analysis}
}
\description{
Analogous to \code{\link{lrgprApply}}, but fits standard (i.e. fixed effects) linear or logistic models for many markers
}
\examples{
# Generate data
n = 100
p = 500
X = matrix(sample(0:2, n*p, replace=TRUE), nrow=n)
y = rnorm(n)
sex = as.factor(sample(1:2, n, replace=TRUE))

# Fit model for all markers
pValues = glmApply( y ~ sex + sex:SNP, features=X, terms=c(3,4))


# Multivariate model
n = 100
p = 1000
m = 10

Y = matrix(rnorm(n*m), nrow=n, ncol=m)
X = matrix(rnorm(n*p), nrow=n, ncol=p)

res = glmApply( Y ~ SNP, features = X, terms=2)

# p-values for univariate hypothesis test of each feature against
# 	each response
res$pValues
}

