% Generated by roxygen2 (4.0.0): do not edit by hand
\name{loss.lrgpr}
\alias{loss.lrgpr}
\title{Loss function}
\usage{
loss.lrgpr(y, yhat, family)
}
\arguments{
\item{y}{observed response}

\item{yhat}{fitted response}

\item{family}{"gaussian" or "binomial"}
}
\description{
Compare observed and fitted response under some loss function
}

