\name{plot.criterion.lrgpr}
\alias{plot.criterion.lrgpr}
\title{Plot AIC/BIC/GCV values for lrgpr() model as rank changes}
\usage{
  plot.criterion.lrgpr(x, col = rainbow(3), ...)
}
\arguments{
  \item{x}{list returned by \code{\link{criterion.lrgpr}}}

  \item{col}{array of 3 colors}

  \item{...}{other arguments}
}
\description{
  Plots the criteria metrics returned by
  \code{\link{criterion.lrgpr}}
}
\seealso{
  \code{\link{criterion.lrgpr}}
}

